+++ engine/src/backend/gl/GLRenderBackend.Frame.cpp
@@ -10,7 +10,6 @@
 #include "engine/scene/Scene.h"
 
 #include <glad/gl.h>
-#include <GLFW/glfw3.h>
 #include <glm/gtc/matrix_transform.hpp>
 
 #include <algorithm>